ご指摘箇所の修正

+++ app/controllers/items_controller.rb
@@ -1,6 +1,7 @@
 class ItemsController < ApplicationController
-  before_action :authenticate_user!, except: [:index,:show,:destroy]
+  before_action :authenticate_user!, except: [:index,:show]
   before_action :set_item, only: [:edit, :show,:update]
+  
  def index
   @items = Item.all.order(created_at: :desc)
  end 
@@ -23,7 +24,7 @@ class ItemsController < ApplicationController
  
  def edit
   unless current_user.id == @item.user.id
-    redirect_to root_path
+    redirect_to root_path 
   end
  end
  
@@ -37,8 +38,10 @@ class ItemsController < ApplicationController
  
  def destroy
   item = Item.find(params[:id])
-  item.destroy
-  redirect_to root_path
+  if item.user_id == current_user.id
+   item.destroy
+   redirect_to root_path
+  end
  end 
 
  
@@ -51,4 +54,5 @@ class ItemsController < ApplicationController
  def set_item
   @item = Item.find(params[:id])
  end
+ 
 end
